Move reading's simple reads into the queries package
chapter content and the two chapter-prereading reads all render a single
aggregate through one mapper, and every finder they use is shared with a
command, so a full port would remove nothing. Apply the ADR's halfway
option: relocate them to queries/ unchanged so use_cases/ can finish as
commands-only, and leave the internals for the day a view outgrows them.
