+++ backend/src/infrastructure/reading/routers/chapter_prereading.py
@@ -5,14 +5,14 @@ from typing import Annotated
 from fastapi import APIRouter, Depends
 from starlette import status
 
-from src.application.reading.use_cases.chapter_prereading.generate_chapter_prereading_use_case import (
-    GenerateChapterPrereadingUseCase,
-)
-from src.application.reading.use_cases.chapter_prereading.get_book_prereading_use_case import (
+from src.application.reading.queries.get_book_prereading_use_case import (
     GetBookPrereadingUseCase,
 )
-from src.application.reading.use_cases.chapter_prereading.get_chapter_prereading_use_case import (
+from src.application.reading.queries.get_chapter_prereading_use_case import (
     GetChapterPrereadingUseCase,
+)
+from src.application.reading.use_cases.chapter_prereading.generate_chapter_prereading_use_case import (
+    GenerateChapterPrereadingUseCase,
 )
 from src.application.reading.use_cases.chapter_prereading.update_prereading_answers_use_case import (
     UpdatePrereadingAnswersUseCase,
